Rearranged actions and permissions
Check in/out actions now were rearranged so everything is more Rails now.
All resources are translated, a description and a manual is written.
There is a little nice hourglass icon for the active entry now.
Last but not least there is a permission for adding entries to the time
tracker and it honours the right of making time log entries when clicking on export
or destroy now.