Added issue delete callback
Deleting an issue which is in the MTT resulted in a "page not found error" when navigating to MTT because of a broken link. Now there is a callback which deletes each line from the MTT which belongs to the deleted issue. The time booked on that issues will not be exported and therefore be lost.